+++ src/app.controller.spec.ts
@@ -15,8 +15,8 @@ describe('AppController', () => {
   });
 
   describe('root', () => {
-    it('should return "Hello World!"', () => {
-      expect(appController.getHello()).toBe('Hello World!');
+    it('should return "Healthcheck Ok"', () => {
+      expect(appController.getHealthcheck()).toBe('Healthcheck Ok');
     });
   });
 });
